+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tag-util",
-  "version": "0.13.49",
+  "version": "0.13.50",
   "description": "Functional FLAC metadata editor with Svelte 5 TagState",
   "type": "module",
   "main": "./out/main/index.js",
