+++ src/renderer/src/components/ui/BadgeField.svelte
