+++ src/renderer/src/components/ui/DropZoneOverlay.svelte
