+++ src/renderer/src/infrastructure/adapters/file-drop-adapter.ts
