+++ src/renderer/src/infrastructure/adapters/path-adapter.test.ts
