+++ src/renderer/src/infrastructure/adapters/path-adapter.ts
