+++ src/renderer/src/infrastructure/adapters/platform-adapter.test.ts
