+++ src/renderer/src/infrastructure/adapters/platform-adapter.ts
